Now the graphics.cpp renders geometry
On the constructor I implemented the VertexBuffer, VertexShader and PixelShader creation. Added a renderInput Layout, created a Struct for the Vertex with the position and color, added the Pixel and Vertex Shader to the project. Discovered that the D3DX11CompileFromFile compiles using the Project path and not the cpp path.