Added the constant buffer
In the graphics.cpp the constant buffer was added from the tutorial 04 but I found a problem when I wanted to access the cb0_values in the window.cpp. I tried to set the cb0_values in static so I only one instance was made without doing copies but I was getting a linking error when using it.